refactor: merge PullRequestAttachmentService.Download test into pullrequest_test.go
